Bot-Envia-Mensagens/requirements.txt
mysql-connector-python
selenium
pyperclip
